+++ components/ControlElements/plusBox.tsx
@@ -1,7 +1,8 @@
 import React from "react";
 import styled, { css } from "styled-components";
 import { MultiLanguage } from "dvote-js";
-import { PLUS_ICON, MINUS_ICON } from "../lib/constants";
+
+import { PLUS_ICON, MINUS_ICON } from "../../lib/constants";
 
 const PlusBoxContainer = styled.div<{ remove?: boolean; add?: boolean }>`
   display: flex;
